+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Rebase on COPASI release 4.28 sources.
 * Add support for the time course automatic interval option.
+* Fix definition of experimental data with columns of type 'ignore'.
 
 # CoRC 0.7.1 (2020-05-24)